Allow to switch back to the old behavior (just in case)
